backend/requirements.txt
flask==2.3.3
openai==1.3.0
json-repair==0.40.0
reportlab==4.0.4
python-pptx==1.0.0
gunicorn==21.2.0
flask-cors==5.0.1
httpx==0.24.1
jsonschema==4.19.1
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
